Move resource mapping into shared module, and add monitored resource labels to trace spans.
