+++ internal/resourcemapping/go.mod
@@ -1,0 +1,5 @@
+module github.com/GoogleCloudPlatform/opentelemetry-operations-go/internal/resourcemapping
+
+go 1.17
+
+require go.opentelemetry.io/collector/model v0.49.0 // indirect
